Analog MPPT
Current sensing and voltage sensing circuits that go into the TL494 to change the PWM signal that will go into the buck converter. This schematic shows how to change the oscillation frequency of the TL494 and also provides a Voltage Protection to shut down all outputs of the TL494 in case of voltage spike.